pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "lag-cli"
requires-python = ">=3.10,<3.14"
authors = [{name = "Lamin Labs", email = "open-source@lamin.ai"}]
readme = "README.md"
dynamic = ["version", "description"]
dependencies = [
    "click",
    "lamindb",
    "nbformat",
    "python-dotenv",
    "requests",
]

[project.scripts]
lag-cli = "lag_cli.__main__:main"

[project.urls]
Home = "https://github.com/laminlabs/lag-cli"

[project.optional-dependencies]
dev = [
    "pre-commit",
    "nox",
    "pytest>=6.0",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore:Jupyter is migrating its paths to use standard platformdirs:DeprecationWarning"
]

[tool.ruff]
src = ["lag_cli", "tests"]
line-length = 88
lint.select = [
    "F",  # Errors detected by Pyflakes
    "E",  # Error detected by Pycodestyle
    "W",  # Warning detected by Pycodestyle
    "I",  # isort
    "D",  # pydocstyle
    "B",  # flake8-bugbear
    "TID",  # flake8-tidy-imports
    "C4",  # flake8-comprehensions
    "BLE",  # flake8-blind-except
    "UP",  # pyupgrade
    "RUF100",  # Report unused noqa directives
    "TCH",  # Typing imports
    "NPY",  # Numpy specific rules
    "PTH",  # Use pathlib
    "S"  # Security
]
lint.ignore = [
    # Do not catch blind exception: `Exception`
    "BLE001",
    # Errors from function calls in argument defaults. These are fine when the result is immutable.
    "B008",
    # line too long -> we accept long comment lines; black gets rid of long code lines
    "E501",
    # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
    "E731",
    # allow I, O, l as variable names -> I is the identity matrix
    "E741",
    # Missing docstring in public module
    "D100",
    # undocumented-public-class
    "D101",
    # Missing docstring in public method
    "D102",
    # Missing docstring in public function
    "D103",
    # Missing docstring in public package
    "D104",
    # __magic__ methods are are often self-explanatory, allow missing docstrings
    "D105",
    # Missing docstring in public nested class
    "D106",
    # Missing docstring in __init__
    "D107",
    ## Disable one in each pair of mutually incompatible rules
    # We don’t want a blank line before a class docstring
    "D203",
    # 1 blank line required after class docstring
    "D204",
    # first line should end with a period [Bug: doesn't work with single-line docstrings]
    # We want docstrings to start immediately after the opening triple quote
    "D213",
    # Section underline is over-indented ("{name}")
    "D215",
    # First line should end with a period
    "D400",
    # First line should be in imperative mood; try rephrasing
    "D401",
    # First word of the first line should be capitalized: {} -> {}
    "D403",
    # First word of the docstring should not be "This"
    "D404",
    # Section name should end with a newline ("{name}")
    "D406",
    # Missing dashed underline after section ("{name}")
    "D407",
    # Section underline should be in the line following the section's name ("{name}")
    "D408",
    # Section underline should match the length of its name ("{name}")
    "D409",
    # No blank lines allowed between a section header and its content ("{name}")
    "D412",
    # Missing blank line after last section ("{name}")
    "D413",
    # camcelcase imported as lowercase
    "N813",
    # module import not at top level of file
    "E402",
    # open()` should be replaced by `Path.open()
    "PTH123",
    # subprocess` call: check for execution of untrusted input - https://github.com/PyCQA/bandit/issues/333
    "S603",
    # Starting a process with a partial executable path
    "S607",
    # asserts
    "S101"
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"docs/*" = ["I", "S101"]
"tests/**/*.py" = [
    "D",  # docstrings are allowed to look a bit off
    "S101", # asserts allowed in tests...
    "ARG", # Unused function args -> fixtures nevertheless are functionally relevant...
    "FBT", # Don't care about booleans as positional arguments in tests, e.g. via @pytest.mark.parametrize()
    "PLR2004", # Magic value used in comparison, ...
    "S311", # Standard pseudo-random generators are not suitable for cryptographic purposes
]
"*/__init__.py" = ["F401"]
